pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nostr-publish"
version = "0.1.0"
description = "Markdown to Nostr long-form publisher with NIP-46 remote signing"
readme = "README.md"
license = "GPL-3.0-or-later"
authors = [
    {name = "Markus Rother", email = "mail@markusrother.de"}
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Editors :: Emacs",
]
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-timeout>=2.1",
    "hypothesis>=6.0",
    "ruff>=0.4.0",
]

[project.scripts]
nostr-publish = "nostr_publish.cli:main"

[project.urls]
Homepage = "https://github.com/941design/emacs-nostr-publish"
Repository = "https://github.com/941design/emacs-nostr-publish"
Issues = "https://github.com/941design/emacs-nostr-publish/issues"

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.ruff.lint.isort]
known-first-party = ["nostr_publish"]
split-on-trailing-comma = false
